prelude-leadgen/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic[email]>=2.5.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
typing-extensions>=4.8.0
python-multipart>=0.0.6
aiofiles>=23.2.1
aiohttp>=3.9.0
fake-useragent>=1.4.0
lxml>=4.9.0
openai>=1.3.0
anthropic>=0.40.0
google-generativeai>=0.3.0
google-genai>=0.3.2
selenium>=4.0.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
-e ../prelude-shared
googlemaps>=4.10.0
PyJWT>=2.8.0
httpx>=0.28.1
psutil>=5.9.0
email-validator>=2.0.0
jinja2>=3.1.0
chardet>=5.2.0

# Optimization dependencies
redis>=5.0.0
circuitbreaker>=1.4.0
asyncio-throttle>=1.0.2

# CSV Mapping Library (shared)
-e ../prelude-csv
sentry-sdk[fastapi]
